FileBuffer_Read: Avoid crash by using size_t instead of unsigned int
Reported-by: John Paul Adrian Glaubitz <glaubitz@physik.fu-berlin.de>
See: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=701680

+++ djmount/file_buffer.c
@@ -212,7 +212,7 @@ FileBuffer_Read (FileBuffer* file, char* buffer,
 		 * to return the exact number of bytes requested.
 		 */
 		do {
-			unsigned int read_size = size - n;
+			size_t read_size = size - n;
 			if (n > 0) {
 				Log_Printf (LOG_DEBUG, 
 					    "UpnpReadHttpGet loop ! url '%s' "
